backend/requirements.txt
flask
flask_cors
pybaseball
python-dotenv
psycopg[binary]
bcrypt
rapidfuzz
scikit-learn
pytest